requirements.txt
bs4==0.0.1
Flask-And-Redis==1.0.0
redis==3.5.3